Enable multiple implementations for Rastreio API
