+++ src/main/java/rastreio/Magic.java
@@ -14,5 +14,15 @@ class Magic {
     0x74, 0x2E, 0x63, 0x66, 0x6D
   }, StandardCharsets.UTF_8);
 
+  public final static String URL2 = new String(new byte[] {
+    0x68, 0x74, 0x74, 0x70, 0x73, 0x3a, 0x2f, 0x2f, 0x77, 0x77,
+    0x77, 0x32, 0x2e, 0x63, 0x6f, 0x72, 0x72, 0x65, 0x69, 0x6f,
+    0x73, 0x2e, 0x63, 0x6f, 0x6d, 0x2e, 0x62, 0x72, 0x2f, 0x73,
+    0x69, 0x73, 0x74, 0x65, 0x6d, 0x61, 0x73, 0x2f, 0x72, 0x61,
+    0x73, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x65, 0x6e, 0x74, 0x6f,
+    0x2f, 0x72, 0x65, 0x73, 0x75, 0x6c, 0x74, 0x61, 0x64, 0x6f,
+    0x2e, 0x63, 0x66, 0x6d
+  }, StandardCharsets.UTF_8);
+
   private Magic() {}
 }
