+++ src/test/java/rastreio/HijackMagicTest.java
@@ -31,4 +31,31 @@ public class HijackMagicTest {
       fail(e.getMessage());
     }
   }
+
+  @Test
+  public void testChangeUrl2() {
+    try {
+      assertEquals("https://www2.correios.com.br/sistemas/rastreamento/resultado.cfm", Magic.URL2);
+
+      HijackMagic.changeUrl2("newUrl");
+      assertEquals("newUrl", Magic.URL2);
+
+      HijackMagic.revertUrl2();
+      assertEquals("https://www2.correios.com.br/sistemas/rastreamento/resultado.cfm", Magic.URL2);
+    } catch (Exception e) {
+      fail(e.getMessage());
+    }
+  }
+
+  @Test
+  public void testRevertUrl2() {
+    try {
+      assertEquals("https://www2.correios.com.br/sistemas/rastreamento/resultado.cfm", Magic.URL2);
+
+      HijackMagic.revertUrl2();
+      assertEquals("https://www2.correios.com.br/sistemas/rastreamento/resultado.cfm", Magic.URL2);
+    } catch (Exception e) {
+      fail(e.getMessage());
+    }
+  }
 }
